pyproject.toml
[tool.poetry]
name = "workout-api"
version = "0.1.0"
description = ""
authors = ["melo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
sqlalchemy = "^2.0.20"
mypy = "^1.5.1"
uvicorn = "^0.23.2"
fastapi = "^0.103.1"
pytest = "^7.4.2"
python-dotenv = "^1.0.0"
pre-commit = "^3.4.0"
httpx = "^0.25.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
